Crate devtools_traits

This module contains shared types and messages for use by devtools/script. The traits are here instead of in script so that the devtools crate can be modified independently of the rest of Servo.

Since these types can be sent through the IPC channel and use non self-describing serializers, the flatten, skip*, tag and untagged serde annotations are not supported. Types like serde_json::Value aren’t supported either. For JSON serialization it is preferred to use a wrapper struct in the devtools crate instead.

Structs

AttrInfo

AttrModification

AutoMargins

CSSError

ComputedNodeLayout

The properties of a DOM node as computed by layout.

ConsoleMessage

ConsoleMessageFields

CssDatabaseProperty

DebuggerNumberVisitor 🔒

DevtoolsPageInfo

EnvironmentInfo

EvaluateJSReply

EventListenerInfo

FrameInfo

FrameOffset

FunctionPreview

HttpRequest

HttpResponse

MatchedRule

NodeInfo

NodeStyle

ObjectPreview

PageError

PauseReason

PropertyDescriptor

https://searchfox.org/mozilla-central/source/devtools/server/actors/object/property-iterator.js#51

RecommendedBreakpointLocation

RuleModification

SecurityInfoUpdate

SourceInfo

StackFrame

StartedTimelineMarker

StyleSheetInfo

TimelineMarker

WorkerId

Enums

AncestorData

BlackboxCoverage

ChromeToDevtoolsControlMsg

Events that the devtools server must act upon.

ConsoleLogLevel

The severity level of a message logged by page content.

DebuggerValue

DevtoolScriptControlMsg

Messages to process in a particular script thread, as instructed by a devtools client. TODO: better error handling, e.g. if pipeline id lookup fails?

DevtoolsControlMsg

Messages to instruct the devtools server to update its known actors/state according to changes in the browser.

DomMutation

GetEnvironmentRequest

GetHTMLType

NavigationState

The state of a page navigation.

NetworkEvent

ScriptToDevtoolsControlMsg

Events that the devtools server must act upon.

ShadowRootMode

TimelineMarkerType

Functions

deserialize_debugger_number 🔒

get_time_stamp