Enum devtools_traits::DevtoolScriptControlMsg

pub enum DevtoolScriptControlMsg {
    EvaluateJS(PipelineId, String, IpcSender<EvaluateJSReply>),
    GetRootNode(PipelineId, IpcSender<Option<NodeInfo>>),
    GetDocumentElement(PipelineId, IpcSender<Option<NodeInfo>>),
    GetChildren(PipelineId, String, IpcSender<Option<Vec<NodeInfo>>>),
    GetLayout(PipelineId, String, IpcSender<Option<ComputedNodeLayout>>),
    ModifyAttribute(PipelineId, String, Vec<Modification>),
    WantsLiveNotifications(PipelineId, bool),
    SetTimelineMarkers(PipelineId, Vec<TimelineMarkerType>, IpcSender<Option<TimelineMarker>>),
    DropTimelineMarkers(PipelineId, Vec<TimelineMarkerType>),
    RequestAnimationFrame(PipelineId, String),
    Reload(PipelineId),
}

Messages to process in a particular script thread, as instructed by a devtools client. TODO: better error handling, e.g. if pipeline id lookup fails?

Variants

EvaluateJS

Evaluate a JS snippet in the context of the global for the given pipeline.

Tuple Fields of EvaluateJS

0: PipelineId1: String2: IpcSender<EvaluateJSReply>

GetRootNode

Retrieve the details of the root node (ie. the document) for the given pipeline.

Tuple Fields of GetRootNode

0: PipelineId1: IpcSender<Option<NodeInfo>>

GetDocumentElement

Retrieve the details of the document element for the given pipeline.

Tuple Fields of GetDocumentElement

0: PipelineId1: IpcSender<Option<NodeInfo>>

GetChildren

Retrieve the details of the child nodes of the given node in the given pipeline.

Tuple Fields of GetChildren

0: PipelineId1: String2: IpcSender<Option<Vec<NodeInfo>>>

GetLayout

Retrieve the computed layout properties of the given node in the given pipeline.

Tuple Fields of GetLayout

0: PipelineId1: String2: IpcSender<Option<ComputedNodeLayout>>

ModifyAttribute

Update a given node’s attributes with a list of modifications.

Tuple Fields of ModifyAttribute

0: PipelineId1: String2: Vec<Modification>

WantsLiveNotifications

Request live console messages for a given pipeline (true if desired, false otherwise).

Tuple Fields of WantsLiveNotifications

0: PipelineId1: bool

SetTimelineMarkers

Request live notifications for a given set of timeline events for a given pipeline.

Tuple Fields of SetTimelineMarkers

0: PipelineId1: Vec<TimelineMarkerType>2: IpcSender<Option<TimelineMarker>>

DropTimelineMarkers

Withdraw request for live timeline notifications for a given pipeline.

Tuple Fields of DropTimelineMarkers

0: PipelineId1: Vec<TimelineMarkerType>

RequestAnimationFrame

Request a callback directed at the given actor name from the next animation frame executed in the given pipeline.

Tuple Fields of RequestAnimationFrame

0: PipelineId1: String

Reload

Direct the given pipeline to reload the current page.

Tuple Fields of Reload

0: PipelineId