Enum devtools_traits::ScriptToDevtoolsControlMsg

pub enum ScriptToDevtoolsControlMsg {
    NewGlobal((PipelineId, Option<WorkerId>), IpcSender<DevtoolScriptControlMsg>, DevtoolsPageInfo),
    ConsoleAPI(PipelineId, ConsoleMessage, Option<WorkerId>),
    FramerateTick(String, f64),
    ReportCSSError(PipelineId, CSSError),
}

Events that the devtools server must act upon.

Variants
NewGlobal((PipelineId, Option<WorkerId>), IpcSender<DevtoolScriptControlMsg>, DevtoolsPageInfo)

A new global object was created, associated with a particular pipeline. The means of communicating directly with it are provided.

ConsoleAPI(PipelineId, ConsoleMessage, Option<WorkerId>)

A particular page has invoked the console API.

FramerateTick(String, f64)

An animation frame with the given timestamp was processed in a script thread. The actor with the provided name should be notified.

ReportCSSError(PipelineId, CSSError)

Report a CSS parse error for the given pipeline