Crate devtools_traits

This module contains shared types and messages for use by devtools/script. The traits are here instead of in script so that the devtools crate can be modified independently of the rest of Servo.

Structs

AttrInfo

AutoMargins

CSSError

CachedConsoleMessageTypes

ComputedNodeLayout
The properties of a DOM node as computed by layout.

ConsoleAPI

ConsoleMessage

DevtoolsPageInfo

HttpRequest

HttpResponse

Modification

NodeInfo

PageError

PreciseTime
A replacement for time::PreciseTime that isn't opaque, so we can serialize it.

StartedTimelineMarker

TimelineMarker

WorkerId

Enums

CachedConsoleMessage

ChromeToDevtoolsControlMsg
Events that the devtools server must act upon.

DevtoolScriptControlMsg
Messages to process in a particular script thread, as instructed by a devtools client.

DevtoolsControlMsg
Messages to instruct the devtools server to update its known actors/state according to changes in the browser.

EvaluateJSReply
Serialized JS return values TODO: generalize this beyond the EvaluateJS message?

LogLevel

NetworkEvent

ScriptToDevtoolsControlMsg
Events that the devtools server must act upon.

TimelineMarkerType

Constants

CONSOLE_API

PAGE_ERROR

_IMPL_DESERIALIZE_FOR_AttrInfo

_IMPL_DESERIALIZE_FOR_AutoMargins

_IMPL_DESERIALIZE_FOR_CSSError

_IMPL_DESERIALIZE_FOR_CachedConsoleMessage

_IMPL_DESERIALIZE_FOR_CachedConsoleMessageTypes

_IMPL_DESERIALIZE_FOR_ComputedNodeLayout

_IMPL_DESERIALIZE_FOR_ConsoleAPI

_IMPL_DESERIALIZE_FOR_ConsoleMessage

_IMPL_DESERIALIZE_FOR_DevtoolScriptControlMsg

_IMPL_DESERIALIZE_FOR_DevtoolsPageInfo

_IMPL_DESERIALIZE_FOR_EvaluateJSReply

_IMPL_DESERIALIZE_FOR_LogLevel

_IMPL_DESERIALIZE_FOR_Modification

_IMPL_DESERIALIZE_FOR_NodeInfo

_IMPL_DESERIALIZE_FOR_PageError

_IMPL_DESERIALIZE_FOR_PreciseTime

_IMPL_DESERIALIZE_FOR_ScriptToDevtoolsControlMsg

_IMPL_DESERIALIZE_FOR_TimelineMarker

_IMPL_DESERIALIZE_FOR_TimelineMarkerType

_IMPL_DESERIALIZE_FOR_WorkerId

_IMPL_SERIALIZE_FOR_AttrInfo

_IMPL_SERIALIZE_FOR_AutoMargins

_IMPL_SERIALIZE_FOR_CSSError

_IMPL_SERIALIZE_FOR_CachedConsoleMessage

_IMPL_SERIALIZE_FOR_CachedConsoleMessageTypes

_IMPL_SERIALIZE_FOR_ComputedNodeLayout

_IMPL_SERIALIZE_FOR_ConsoleAPI

_IMPL_SERIALIZE_FOR_ConsoleMessage

_IMPL_SERIALIZE_FOR_DevtoolScriptControlMsg

_IMPL_SERIALIZE_FOR_DevtoolsPageInfo

_IMPL_SERIALIZE_FOR_EvaluateJSReply

_IMPL_SERIALIZE_FOR_LogLevel

_IMPL_SERIALIZE_FOR_Modification

_IMPL_SERIALIZE_FOR_NodeInfo

_IMPL_SERIALIZE_FOR_PageError

_IMPL_SERIALIZE_FOR_PreciseTime

_IMPL_SERIALIZE_FOR_ScriptToDevtoolsControlMsg

_IMPL_SERIALIZE_FOR_TimelineMarker

_IMPL_SERIALIZE_FOR_TimelineMarkerType

_IMPL_SERIALIZE_FOR_WorkerId