Crate devtools_traits

This module contains shared types and messages for use by devtools/script. The traits are here instead of in script so that the devtools crate can be modified independently of the rest of Servo.

Structs

AttrInfo

AttrModification

AutoMargins

CSSError

CachedConsoleMessageTypes

ComputedNodeLayout

The properties of a DOM node as computed by layout.

ConsoleLog

ConsoleMessage

CssDatabaseProperty

DevtoolsPageInfo

HttpRequest

HttpResponse

NodeInfo

NodeStyle

PageError

RuleModification

StartedTimelineMarker

TimelineMarker

WorkerId

Enums

CachedConsoleMessage

ChromeToDevtoolsControlMsg

Events that the devtools server must act upon.

DevtoolScriptControlMsg

Messages to process in a particular script thread, as instructed by a devtools client. TODO: better error handling, e.g. if pipeline id lookup fails?

DevtoolsControlMsg

Messages to instruct the devtools server to update its known actors/state according to changes in the browser.

EvaluateJSReply

Serialized JS return values TODO: generalize this beyond the EvaluateJS message?

LogLevel

NavigationState

The state of a page navigation.

NetworkEvent

ScriptToDevtoolsControlMsg

Events that the devtools server must act upon.

TimelineMarkerType