Crate devtools_traits

This module contains shared types and messages for use by devtools/script. The traits are here instead of in script so that the devtools crate can be modified independently of the rest of Servo.

Structs

AttrInfo

DevtoolsPageInfo

Modification

NodeInfo

Enums

ConsoleMessage

DevtoolScriptControlMsg
Messages to process in a particular script task, as instructed by a devtools client.

DevtoolsControlMsg
Messages to the instruct the devtools server to update its known actors/state according to changes in the browser.

EvaluateJSReply
Serialized JS return values TODO: generalize this beyond the EvaluateJS message?

ScriptDevtoolControlMsg
Messages to instruct devtools server to update its state relating to a particular tab.

Type Definitions

DevtoolsControlChan

DevtoolsControlPort