Enum devtools_traits::DevtoolScriptControlMsg

pub enum DevtoolScriptControlMsg {
    EvaluateJS(PipelineId, String, IpcSender<EvaluateJSReply>),
    GetRootNode(PipelineId, IpcSender<NodeInfo>),
    GetDocumentElement(PipelineId, IpcSender<NodeInfo>),
    GetChildren(PipelineId, String, IpcSender<Vec<NodeInfo>>),
    GetLayout(PipelineId, String, IpcSender<ComputedNodeLayout>),
    GetCachedMessages(PipelineId, CachedConsoleMessageTypes, IpcSender<Vec<CachedConsoleMessage>>),
    ModifyAttribute(PipelineId, String, Vec<Modification>),
    WantsLiveNotifications(PipelineId, bool),
    SetTimelineMarkers(PipelineId, Vec<TimelineMarkerType>, IpcSender<TimelineMarker>),
    DropTimelineMarkers(PipelineId, Vec<TimelineMarkerType>),
    RequestAnimationFrame(PipelineId, String),
    Reload(PipelineId),
}

Messages to process in a particular script thread, as instructed by a devtools client.

Variants
EvaluateJS(PipelineId, String, IpcSender<EvaluateJSReply>)

Evaluate a JS snippet in the context of the global for the given pipeline.

GetRootNode(PipelineId, IpcSender<NodeInfo>)

Retrieve the details of the root node (ie. the document) for the given pipeline.

GetDocumentElement(PipelineId, IpcSender<NodeInfo>)

Retrieve the details of the document element for the given pipeline.

GetChildren(PipelineId, String, IpcSender<Vec<NodeInfo>>)

Retrieve the details of the child nodes of the given node in the given pipeline.

GetLayout(PipelineId, String, IpcSender<ComputedNodeLayout>)

Retrieve the computed layout properties of the given node in the given pipeline.

GetCachedMessages(PipelineId, CachedConsoleMessageTypes, IpcSender<Vec<CachedConsoleMessage>>)

Retrieve all stored console messages for the given pipeline.

ModifyAttribute(PipelineId, String, Vec<Modification>)

Update a given node's attributes with a list of modifications.

WantsLiveNotifications(PipelineId, bool)

Request live console messages for a given pipeline (true if desired, false otherwise).

SetTimelineMarkers(PipelineId, Vec<TimelineMarkerType>, IpcSender<TimelineMarker>)

Request live notifications for a given set of timeline events for a given pipeline.

DropTimelineMarkers(PipelineId, Vec<TimelineMarkerType>)

Withdraw request for live timeline notifications for a given pipeline.

RequestAnimationFrame(PipelineId, String)

Request a callback directed at the given actor name from the next animation frame executed in the given pipeline.

Reload(PipelineId)

Direct the given pipeline to reload the current page.